chore: mama-lang now uses only one `translation` file to heighlight and adding new keyword
